Split Task_State into Background_State and Cancellation_State.
These are two references to the same shared task state object.

Background_State is to be used by the background task. It provides
procedures for changing the task state, and testing for cancellation.

Cancellation_State is to be used by the foreground task, which
initiated the background task. It provides procedures for cancelling
the background task.